chores(gateway): move functions to the right module
