+++ apps/gateway/src/services/provider-factory.ts
@@ -8,6 +8,8 @@ import { withCanonicalIdsForVertex } from "@hebo-ai/gateway/providers/vertex";
 import { withCanonicalIdsForVoyage } from "@hebo-ai/gateway/providers/voyage";
 import { createVoyage } from "voyage-ai-provider";
 
+import { getSecret } from "@hebo/shared-api/utils/secrets";
+
 import type {
   ApiKeyProviderConfig,
   BedrockProviderConfig,
@@ -18,6 +20,39 @@ import type {
 import { buildWifOptions } from "./aws-wif";
 
 import type { ProviderV3 } from "@ai-sdk/provider";
+
+export async function loadProviderSecrets() {
+  const [
+    groqApiKey,
+    bedrockRoleArn,
+    bedrockRegion,
+    voyageApiKey,
+    vertexServiceAccountEmail,
+    vertexAudience,
+    vertexLocation,
+    vertexProject,
+  ] = await Promise.all([
+    getSecret("GroqApiKey"),
+    getSecret("BedrockRoleArn"),
+    getSecret("BedrockRegion"),
+    getSecret("VoyageApiKey"),
+    getSecret("VertexServiceAccountEmail"),
+    getSecret("VertexAwsProviderAudience"),
+    getSecret("VertexLocation"),
+    getSecret("VertexProject"),
+  ]);
+
+  return {
+    groqApiKey,
+    bedrockRoleArn,
+    bedrockRegion,
+    voyageApiKey,
+    vertexServiceAccountEmail,
+    vertexAudience,
+    vertexLocation,
+    vertexProject,
+  };
+}
 
 export function createProvider(
   slug: ProviderSlug,
